Move "#create", "#parse", "#parse_from_title" from Identifier::Base to Identifier
