+++ lib/pubid/iso.rb
@@ -14,6 +14,7 @@ require_relative "iso/stage"
 require_relative "iso/type"
 require_relative "iso/harmonized_stage_code"
 require_relative "iso/transformer"
+require_relative "iso/identifier"
 require_relative "iso/identifier/base"
 require_relative "iso/identifier/international_standard"
 require_relative "iso/identifier/international_standardized_profile"
