+++ lib/pubid/iso/transformer.rb
@@ -93,7 +93,7 @@ module Pubid::Iso
 
     rule(joint_document: subtree(:joint_document)) do |context|
       context[:joint_document] =
-        Identifier::Base.create(**context[:joint_document])
+        Identifier.create(**context[:joint_document])
       context
     end
